# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY API",
        "description": "Documentation OpenAPI pour l'API OMPAY. Application de paiement mobile avec authentification par SMS, transferts d'argent et paiements QR.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Environnement local"
        }
    ],
    "paths": {
...
        "/api/compte": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Consulter les informations du compte bancaire",
                "operationId": "7abca07d85bb3033fa6991ad42a51da1",
                "responses": {
                    "200": {
                        "description": "Informations du compte retournées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "compte": {
                                                    "$ref": "#/components/schemas/Compte"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/compte/transfert": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Effectuer un transfert d'argent vers un autre utilisateur",
                "operationId": "7b6e256f0b3c31ff2f4af9a196862b59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "montant": {
                                        "type": "number",
                                        "example": 25
                                    },
                                    "to_compte_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid-here"
                                    },
                                    "to_phone": {
                                        "type": "string",
                                        "example": "771234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert effectué"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors du transfert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/qr/payer": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Payer à l'aide d'un code marchand",
                "operationId": "0f103e155b6669ada424cee17c7fb0ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_marchand"
                                ],
                                "properties": {
                                    "code_marchand": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de paiement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float"
                    },
                    "devise": {
                        "type": "string",
                        "example": "XOF"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}